Crate bcrypt_small

Structs

Hash

A bcrypt hash, comprising a work factor, salt, and digest bytes. Holds information equivalent to the $2b$… format created by OpenBSD’s bcrypt.

Salt

A bcrypt salt.

WorkFactor

A bcrypt work factor.

Enums

CompareError

A bcrypt hashing error.

HashError

A bcrypt hashing error for new hashes.

ParseError

An error parsing a formatted bcrypt hash string.

Constants

FORMATTED_HASH_SIZE

The number of ASCII bytes in a $2b$… formatted bcrypt hash string.

KEY_SIZE_MAX

The maximum number of bytes in a bcrypt key.

Functions

compare

Compares a password to an existing hash.

hash

Creates a new password hash.